-add round cut change function

+++ lib/src/main/java/cyoap_main/design/choice/ChoiceSet.java
@@ -55,7 +55,7 @@ public class ChoiceSet {
     public float height;
     public boolean isClicked = false;
 
-    public float round = 0;
+    public int round = 0;
 
     @JsonIgnore
     public Bound2f bound;
